requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
openai==1.52.2
requests==2.32.3
google-cloud-vision==3.7.4
httpx==0.27.2